[MESURES] Simplifie le métier
...en exposant directement les exigences liées, plutôt qu'un moyen de les récupérer
