+++ back/src/api/mesures/mesurePresentation.ts
@@ -1,10 +1,7 @@
 import { Mesure } from '../../metier/mesure';
-import { EntrepotExigence } from '../../metier/nis2/entrepotExigence';
 
-export const mesurePresentation = async (mesure: Mesure, entrepotExigence: EntrepotExigence) => {
-  const exigences = await mesure.exigences(entrepotExigence);
+export const mesurePresentation = async (mesure: Mesure) => {
   return {
     ...mesure,
-    exigences,
   };
 };
