+++ back/src/infra/entrepotMesurePostgres.ts
@@ -2,6 +2,7 @@ import Knex from 'knex';
 import config from '../../knexfile';
 import { EntrepotMesure } from '../metier/entrepotMesure';
 import { LienPourAllerPlusLoin, Mesure, type Risque } from '../metier/mesure';
+import { EntrepotExigence } from '../metier/nis2/entrepotExigence';
 
 export type MesurePersistee = {
   id: string;
@@ -19,13 +20,13 @@ export type MesurePersistee = {
 export class EntrepotMesurePostgres implements EntrepotMesure {
   knex: Knex.Knex;
 
-  constructor() {
+  constructor(private readonly entrepotExigence: EntrepotExigence) {
     this.knex = Knex(config);
   }
 
   async tous(): Promise<Mesure[]> {
     const mesuresLues = await this.knex<MesurePersistee>('mesures');
-    return mesuresLues.map((mesure) => this.convertisEnMesure(mesure));
+    return Promise.all(mesuresLues.map((mesure) => this.convertisEnMesure(mesure)));
   }
 
   async parId(id: string): Promise<Mesure | undefined> {
@@ -33,7 +34,27 @@ export class EntrepotMesurePostgres implements EntrepotMesure {
     return mesurePersistee ? this.convertisEnMesure(mesurePersistee) : undefined;
   }
 
-  private convertisEnMesure(mesurePersistee: MesurePersistee): Mesure {
+  private async convertisEnMesure(mesurePersistee: MesurePersistee): Promise<Mesure> {
+    const toutesLesExigencesNIS2 = await this.entrepotExigence.parReferentiel('NIS2');
+
+    const exigences = mesurePersistee.references_nis2
+      .map((ref) => {
+        const exigenceTrouvee = toutesLesExigencesNIS2.find((e) => e.reference === ref);
+        if (!exigenceTrouvee) {
+          return null;
+        }
+        return {
+          contenu: exigenceTrouvee.contenu,
+          contenuEnAnglais: exigenceTrouvee.contenuEnAnglais,
+          reference: exigenceTrouvee.reference,
+          thematique: exigenceTrouvee.thematique,
+          objectifSecurite: exigenceTrouvee.objectifSecurite,
+          entitesCible: exigenceTrouvee.entitesCible,
+          correspondances: {},
+        };
+      })
+      .filter((e) => !!e);
+
     return new Mesure(
       mesurePersistee.id,
       mesurePersistee.titre,
@@ -44,7 +65,7 @@ export class EntrepotMesurePostgres implements EntrepotMesure {
       mesurePersistee.ordre,
       mesurePersistee.risques,
       mesurePersistee.liens,
-      mesurePersistee.references_nis2
+      exigences
     );
   }
 }
